duck_tracker/apps/ducks/utils.py
from import_export.formats import base_formats


def get_default_formats():
    """
    Returns available export formats.
    """
    formats = (
        base_formats.CSV,
        # base_formats.XLS,
        # base_formats.XLSX,
        # base_formats.TSV,
        # base_formats.ODS,
        # base_formats.JSON,
        # base_formats.YAML,
        # base_formats.HTML,
    )
    return [f for f in formats if f().can_export()]
